Fixed a bug where corridor walls weren't being added to list
Also, changed algorithm so that instead of choosing a random wall from a
list of walls, we store each feature's (room/corridor) walls in a
seperate list, so instead we have a list of lists. This is because I was
seeing a lot of corridors with dead ends and figured that since rooms
have a lot more wall tiles than corridors, it was much more likely to
pick a room to build features from rather than a corridor. Instead we
now store the list of walls added by a feature as another list within
the walls list (append vs. extend), and then choose a random list before
choosing a random wall from that list. This gives rooms/corridors
an equal chance since that first random choice is not dependent on how
many wall tiles the feature has.