Added throwable weapons now, made changes to equipping
Added throwable weapons to the HUD, as well as adding item quantities.
Weapons can now be thrown if they are throwable, although the interface
to equip things is not clear yet. I think I will change the equipping
interface at some point to ask the player which slot they want to equip
the selected item in, instead of equipping in a default slot.